+++ rest/test/db/utils/testDbOptions.js
@@ -28,6 +28,6 @@ module.exports = {
 		const mongoHost = args.mongoHost || '127.0.0.1';
 		return `mongodb://${mongoHost}:27017/`;
 	})(),
-	networkId: catapult.model.networkInfo.networks.mijinTest.id,
+	networkId: catapult.model.networkInfo.networks.testnet.id,
 	connectionPoolSize: 5
 };